DA#229: Modified: Move openDocument to new folder
